feat: Linting and Typscript issues fixed, added Notifications
Fixed issues that could be found with liniting and the command npx tsc.
Most of the issues where unused variables.
added the expo Notifications package and used it to notify the user at
currently hardcoded times to log food if there was no food logged for
the certain meal after a point in daytime (breakfast,lunch,dinner ~
6am,12am,6pm)